pyproject.toml
[tool.poetry]
name = "strawberry-django-extras"
version = "0.1.9"
description = "Strawberry GraphQL Django Extras"
authors = ["Giorgos Gernas <m4riok@gmail.com>"]
maintainers = ["Giorgos Gernas <m4riok@gmail.com>"]
repository = "https://github.com/m4riok/strawberry-django-extras"
license = "MIT"
readme = "README.md"
keywords = ["graphql", "api", "django", "strawberry-graphql"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
Django = ">=4.0"
strawberry-graphql = ">=0.215.1"
strawberry-graphql-django = ">=0.25.0"
makefun = ">=1.15.2"
PyJWT = ">=2.1.0"
cryptography = ">=38.0"
graphql-core = ">=3.1.5"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
pytest = "^7.1.2"
pytest-asyncio = "^0.21.0"
pytest-django = "^4.1.0"
pytest-pythonpath = "^0.7.3"
pytest-watch = "^4.2.0"
pytest-mock = "^3.5.1"
pytest-snapshot = "^0.9.0"
pytest-cov = "^4.0.0"
mkdocs = "^1.4.2"
mkdocs-markdownextradata-plugin = "^0.2.5"
mkdocs-material = "^9.0.4"
mkdocs-minify-plugin = "^0.7.1"
pymdown-extensions = "^10.0.1"
mkdocstrings = { version = "^0.23.0", extras = ["python"] }
ruff = "^0.1.6"
django-types = "^0.18.0"
Markdown = "^3.3.7"
Pygments = "^2.15.1"
factory-boy = "^3.2.1"
black = "^23.9.1"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
select = [
    "A",
    "ASYNC100",
    "ASYNC101",
    "ASYNC102",
    "B",
    "BLE",
    "C4",
    "COM",
    "DTZ",
    "E",
    "ERA",
    "EXE",
    "F",
    "FURB",
    "G",
    "I",
    "ICN001",
    "INP",
    "ISC",
    "N",
    "PERF",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "PYI",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "SLOT",
    "T10",
    "T20",
    "TCH",
    "TID",
    "TRY",
    "UP",
    "W",
    "YTT",
]
ignore = [
    "TRY003",
    "PLR6301",
    "PLR6201",
    "PLC0415",
    "W191",
    "E111",
    "E114",
    "E117",
    "D206",
    "D300",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "COM812",
    "COM819",
    "ISC001",
    "ISC002",
    "E501"
]
target-version = "py38"
exclude = [
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    ".venv",
    "__pycached__",
    "_build",
    "buck-out",
    "build",
    "dist",
]
[tool.ruff.format]
exclude = ["**/migrations/*"]
[tool.ruff.per-file-ignores]
"tests/*" = ["A003", "PLW0603", "PLR2004"]
"examples/*" = ["A003"]
"**/migrations/*" = ["RUF012", "E501", "Q000", "I001", "COM812"]
"asp/*" = ["N999"]

[tool.black]
target-version = ['py38', 'py39', 'py310', 'py311']
preview = true
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | __pycached__
  | _build
  | buck-out
  | build
  | dist
)/
'''
